refactor: home 디렉터리 구조 변경
