+++ frontend/rush/src/components/home/DefaultMapPage.js
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from 'react';
 import DefaultMap from "./DefaultMap";
 import {CLIENT_ID} from "../../constants/GoogleMapAuth";
-import Menu from "./button/MenuButton";
+import Menu from "./menu/Menu";
 import LoginButton from "./button/LoginButton";
 import WriteButton from "./button/WriteButton";
 import WindowSize from "../../util/WindowSize";
@@ -15,8 +15,8 @@ import {
 } from "../../api/FindMapArticlesApi";
 import {GROUPED, PRIVATE, PUBLIC} from "../../constants/MapType";
 import {withRouter} from "react-router-dom";
-import CreateGroupModal from "./group/create/CreateGroupModal";
-import JoinGroupModal from "./group/join/JoinGroupModal";
+import CreateGroupModal from "./menu/group/create/CreateGroupModal";
+import JoinGroupModal from "./menu/group/join/JoinGroupModal";
 import MapType from "./MapTypeStyle";
 import findGroupApi from "../../api/FindGroupApi";
 import MyLocationButton from "./button/MyLocationButton";
