+++ frontend/rush/src/components/home/menu/group/create/CompleteButton.js
@@ -1,5 +1,5 @@
 import React from "react";
-import createGroupApi from "../../../../api/CreateGroupApi";
+import createGroupApi from "../../../../../api/CreateGroupApi";
 import {withRouter} from "react-router-dom";
 
 const CompleteButton = ({isCreateGroupModalOpen, groupName, history}) => {
