+++ frontend/rush/src/components/home/menu/group/join/CompleteButton.js
@@ -1,5 +1,5 @@
 import React from "react";
-import joinGroupApi from "../../../../api/JoinGroupApi";
+import joinGroupApi from "../../../../../api/JoinGroupApi";
 
 const CompleteButton = ({isCreateGroupModalOpen, invitationCode, history}) => {
   return <>
