+++ frontend/rush/src/components/home/menu/styled.css
